+++ versioned_value.hh
@@ -6,7 +6,7 @@
 template <typename T, typename=void>
 struct versioned_value_struct /*: public threadinfo::rcu_callback*/ {
   typedef T value_type;
-  typedef uint32_t version_type;
+  typedef uint64_t version_type;
 
   versioned_value_struct() : version_(), value_() {}
